server/requirements.txt
# Server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
websockets>=12.0
sqlalchemy>=2.0.23
alembic>=1.12.1
redis>=5.0.1
celery>=5.3.4
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
